slightly changed the import statements
in case we need to use the packages later. They should still work the same way as before but now can be used better... e.g. if we need to use a class or function inside the packages now it's a lot easier. Feel free to revert if this doesnt work